# Optimize the performance of sum by using universal intrinsics