@martin-frbg 2020-10-14
Oops - that code is single-precision only, right ? arm/sum.c gets built for both SSUM and DSUM (where the -DDOUBLE automagically changes the FLOAT type to double instead of float)